adding files to github

+++ blogPage/views.py
@@ -1,9 +1,11 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from addBlog.models import Blog
+from django.http import JsonResponse
 from .models import Like, Dislike, Comment
 from .forms import CommentForm
 from django.urls import reverse
 from django.utils import timezone
+
 # Create your views here.
 def blog_page(request) :
     
@@ -20,9 +22,6 @@ def blog_page(request) :
             blog = Blog.objects.filter().order_by('added_date')
         else:
             blog = Blog.objects.filter().order_by('-added_date')
-            
-        
-        
         
         return render(request , "blog_page.html" , {'blog' : blog})
     
@@ -31,48 +30,34 @@ def blog_page(request) :
     return render(request , "blog_page.html" , {'blog' : blog}) 
 
 
-def blog_details(request , id) :
+def blog_details(request , id , comment_id = None , error = "" , info = "") :
     blog = Blog.objects.filter(id = id).get()
     likes = Like.objects.filter(blog=blog).count()
     dislikes = Dislike.objects.filter(blog=blog).count()
-    count = likes-dislikes
-    current_date = timezone.now()
-    days = (current_date - blog.added_date).days
+    count = likes - dislikes
+    
+    def getTime(target_date) :
+        current_date = timezone.now()
+        days = (current_date - target_date).days
+        seconds = (current_date - target_date).seconds
+        minutes = seconds // 60
+        hours = minutes // 60
+        time = str(days) + " days ago"
+        if days == 0 :
+            time = str(hours) + " hours ago"
+            if hours == 0 :
+                time = str(minutes) + " minutes ago"
+                if minutes == 0 : 
+                    time = str(seconds) + " seconds ago"
+        return time
+        
     comments = Comment.objects.filter(blog=blog)
-    
-    
-    # Assuming you've already fetched your comments
-    # comment_dict = {}
-
-    # for comment in comments:
-    #     # print(comment.parent_comment)
-    #     if comment.parent_comment is None:
-    #         # This is a top-level comment
-    #         comment_dict[comment] = []
-            
-            
-    # hope = []
-    
-    # for comment in comments :
-    #     level = []
-    #     print(comment.text)
-    #     origin_comment = comment
-    #     while comment.parent_comment != None :
-    #         level.append("hope")
-    #         comment = comment.parent_comment
-    #         if comment == None : break
-    #         print(" " * len(level) + comment.text)
-    #     hope.append({
-    #         "comment" : origin_comment,
-    #         "level" : level
-    #     })
-    
-    
+    comments_count = len(comments)
     mapping_to_comments = {}
     
     adj = {}
     
-    for comment in comments :
+    for comment in comments.order_by('-created_at') :
         parent = comment.parent_comment
         mapping_to_comments[comment.id] = comment
         if parent == None :
@@ -90,46 +75,61 @@ def blog_details(request , id) :
         visited[node] = True
         hope_2.append({
             "comment" : mapping_to_comments[node],
-            "level" : [0 for _ in range(level)]})
+            "level" : [0 for _ in range(level)],
+            "time" : getTime(mapping_to_comments[node].created_at)
+        })
         if node not in adj.keys() :
             return
         for child in adj[node] :
             dfs(child , level + 1)
             
-    for comment in comments :
-        if comment.id not in visited.keys() :
+    for comment in comments.order_by('-created_at') :
+        if comment.id not in visited.keys() and comment.parent_comment == None:
             dfs(comment.id , 0)
     
+    recent_blogs = []
     
-
-    # for comment in comments:
-    #     print(comment)
-    #     if comment.parent_comment is not None and comment.parent_comment in comment_dict:
-    #         # This comment is a reply to an existing comment
-    #         comment_dict[comment.parent_comment].append(comment)
-    # print(comment_dict)
-# Pass comment_dict to the template
-
-    return render(request , "blog_details.html", {'blog' : blog ,'likes' : likes, 'dislikes': dislikes,'count': count,'days':days,'hope' : hope_2})
+    for blog_ in Blog.objects.filter().order_by('-added_date') :
+        recent_blogs.append({
+            "id" : blog_.id,
+            "author" : blog_.author,
+            "title" : blog_.title,
+            "time" : getTime(blog_.added_date)
+        })
+        
+    print("like : " , likes)
+    print("dislike : " , dislikes)
+        
+    return render(request , "blog_details_new.html", {'blog' : blog ,'likes' : likes, 'dislikes': dislikes,'count': count,'time':getTime(blog.added_date),'hope' : hope_2, 'comments_count' : comments_count , 'recent_blogs' : recent_blogs , 'comment_id_focus' : comment_id , 'error' : error , 'info' : info})
 
 def blog_likes(request , id) :
+    # storing like - dislike in likes only.
     user = request.user
     blog = Blog.objects.filter(id = id).get()
     
     current_likes = blog.likes
     liked = Like.objects.filter(user=user,blog=blog).count()
+    disliked = Dislike.objects.filter(user = user , blog = blog).count()
+    
+    if disliked :
+        return blog_details(request, id = id , error = "You already disliked the blog. First remove dislike to update.")
+        
+    print("like 0 : " , current_likes)
     
     if not liked:
         liked = Like.objects.create(user=user,blog=blog)
         current_likes = current_likes + 1
+        info = "Like added"
     else:
         liked = Like.objects.filter(user=user,blog=blog).delete()
         current_likes = current_likes - 1
+        info = "Like removed"
         
     blog.likes = current_likes
+    print("like 1 : " , current_likes)
     blog.save()
     
-    return redirect('blog_details', id=id)
+    return blog_details(request , id=id , info = info)
 
 def blog_dislikes(request , id) :
     user = request.user
@@ -137,18 +137,24 @@ def blog_dislikes(request , id) :
     
     current_dislikes = blog.dislikes
     disliked = Dislike.objects.filter(user=user,blog=blog).count()
+    liked = Like.objects.filter(user=user,blog=blog).count()
+    
+    if liked :
+        return blog_details(request , id = id , error = "You already liked the blog. First remove like to update.")
     
     if not disliked:
         disliked = Dislike.objects.create(user=user,blog=blog)
         current_dislikes = current_dislikes + 1
+        info = "Dislike added"
     else:
         disliked = Dislike.objects.filter(user=user,blog=blog).delete()
         current_dislikes = current_dislikes - 1
+        info = "Dislike removed"
         
     blog.dislikes = current_dislikes
     blog.save()
     
-    return redirect('blog_details', id=id)
+    return blog_details(request , id=id , info = info)
 
 
 
@@ -172,7 +178,7 @@ def add_comment(request, blog_id, comment_id = None):
             comment.user = request.user  # Assuming user is logged in
             comment.parent_comment = parent_comment  # Set the parent comment
             comment.save()
-            return redirect('blog_details', id=blog.id)
+            return redirect('blog_details', id=blog.id , comment_id = comment.id)
 
     # If the request method is not POST or the form is not valid, render the page
     form = CommentForm()
@@ -186,3 +192,9 @@ def add_comment(request, blog_id, comment_id = None):
     })
 
 
+def add_coomment_js(request , blog_id , comment_id) :
+    data = {
+        "value" : "done"
+    }
+    print("data : " , data)
+    return JsonResponse(data)
